+++ Frameworks/AppKit/NSColor.subproj/NSColor_CGColor.h
@@ -17,9 +17,9 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 - initWithColorRef:(CGColorRef)colorRef spaceName:(NSString *)spaceName;
 
 + (NSColor *)colorWithColorRef:(CGColorRef)colorRef spaceName:(NSString *)spaceName;
-+ (NSColor *)colorWithGray:(float)gray alpha:(float)alpha spaceName:(NSString *)spaceName;
-+ (NSColor *)colorWithCyan:(float)cyan magenta:(float)magenta yellow:(float)yellow black:(float)black alpha:(float)alpha spaceName:(NSString *)spaceName;
-+ (NSColor *)colorWithRed:(float)red green:(float)green blue:(float)blue alpha:(float)alpha spaceName:(NSString *)spaceName;
-+ (NSColor *)colorWithHue:(float)hue saturation:(float)saturation brightness:(float)brightness alpha:(float)alpha spaceName:(NSString *)spaceName;
++ (NSColor *)colorWithGray:(CGFloat)gray alpha:(CGFloat)alpha spaceName:(NSString *)spaceName;
++ (NSColor *)colorWithCyan:(CGFloat)cyan magenta:(CGFloat)magenta yellow:(CGFloat)yellow black:(CGFloat)black alpha:(CGFloat)alpha spaceName:(NSString *)spaceName;
++ (NSColor *)colorWithRed:(CGFloat)red green:(CGFloat)green blue:(CGFloat)blue alpha:(CGFloat)alpha spaceName:(NSString *)spaceName;
++ (NSColor *)colorWithHue:(CGFloat)hue saturation:(CGFloat)saturation brightness:(CGFloat)brightness alpha:(CGFloat)alpha spaceName:(NSString *)spaceName;
 
 @end
